[feat] Route 적용, 컴포넌트명 변경
