+++ src/App.js
@@ -1,3 +1,4 @@
+import { Outlet } from "react-router-dom";
 import { Header } from "./layout/Header";
 import { View } from "./layout/View";
 import "./style/css/style.min.css";
