+++ src/components/GNB.js
@@ -1,8 +1,14 @@
+import { NavLink } from "react-router-dom";
+
+function getLinkClass({ isActive }) {
+  return isActive ? "link on" : "link";
+}
+
 export function GNB () {
   return (
     <ul className="gnb">
-      <li className="gnb-list"><a href="" className="link">자유게시판</a></li>
-      <li className="gnb-list"><a href="" className="link">중고마켓</a></li>
+      <li className="gnb-list"><NavLink to="/" className={getLinkClass}>자유게시판</NavLink></li>
+      <li className="gnb-list"><NavLink to="/items" className={getLinkClass}>중고마켓</NavLink></li>
     </ul>
   );
 }
