+++ src/index.js
@@ -1,10 +1,4 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App.js";
+import ReactDOM from 'react-dom';
+import Main from './Main';
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
+ReactDOM.render(<Main />, document.getElementById('root'));
