+++ src/layout/Header.js
@@ -1,5 +1,6 @@
+import { Link } from "react-router-dom";
 import { GNB } from "../components/GNB";
-import { Logo } from "./Logo";
+import { Logo } from "../components/Logo";
 
 export function Header() {
   
@@ -7,7 +8,7 @@ export function Header() {
     <header className="header">
       <div className="header-wrap">
         <div className="section-header">
-          <Logo>판다마켓</Logo>
+          <Link to="/"><Logo>판다마켓</Logo></Link>
           <GNB/>
         </div>
         <div>
