+++ src/page/HomePage.js
@@ -1,0 +1,5 @@
+export function HomePage() {
+  return (
+    <p>메인</p>
+  )
+}
