+++ src/style/scss/_component.scss
@@ -206,6 +206,9 @@
       padding:0 4px;
     }
   }
+  .link.on {
+    color:var(--brand-blue);
+  }
 }
 
 .itemlists {
